requirements.txt
# Core AI and LangChain
langchain==0.3.14
langchain-core==0.3.29
langchain-community==0.3.14
langchain-experimental==0.3.4
langchain-google-genai==2.0.8
google-generativeai>=0.3.0

# FastAPI and Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# LINE Bot SDK (optional)
line-bot-sdk>=3.5.0

# Data Validation and Configuration
pydantic==2.10.5
pydantic-settings>=2.0.0
python-dotenv==1.0.1

# Redis for Memory Management
redis>=5.0.0

# HTTP Clients
aiohttp>=3.9.0
httpx>=0.25.0

# Web Scraping and Search
beautifulsoup4==4.12.3
duckduckgo-search==7.2.1
html2text==2024.2.26
google-search-results>=2.4.2  # SerpAPI

# Kubernetes Client
kubernetes>=28.0.0

# Console and UI
pyfiglet==1.0.2
pyreadline3==3.5.4; sys_platform == 'win32'
rich==13.9.4

# Utilities
python-multipart>=0.0.6

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
